Lab03-Symmetric-Encryption/Program2/requirements.txt
appnope==0.1.2
backcall==0.2.0
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.2
jupyter-core==4.7.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycrypto==2.6.1
pycryptodome==3.11.0
Pygments==2.10.0
python-dateutil==2.8.2
pyzmq==22.3.0
six==1.16.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
